+++ includes/plugins/jetpack/class-jetpack-custom-post-types.php
@@ -8,7 +8,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.2.0
+ * @version  2.5.0
  *
  * Contents:
  *
@@ -35,13 +35,15 @@ class Modern_Jetpack_Custom_Post_Types {
 		 * Constructor
 		 *
 		 * @since    2.0.0
-		 * @version  2.0.2
+		 * @version  2.5.0
 		 */
 		private function __construct() {
 
 			// Requirements check
 
-				if ( ! Jetpack::is_active() && ! Jetpack::is_development_mode() ) {
+				$status = new Automattic\Jetpack\Status();
+
+				if ( ! Jetpack::is_active() && ! $status->is_offline_mode() ) {
 					return;
 				}