+++ template-parts/menu/menu-primary.php
@@ -10,32 +10,45 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.2.0
+ * @version  2.5.0
  */
 
-
-
-
-
-// Helper variables
-
-	$is_mobile_nav_enabled = Modern_Library_Customize::get_theme_mod( 'navigation_mobile' );
-
+$is_mobile_nav_enabled = Modern_Library_Customize::get_theme_mod( 'navigation_mobile' );
 
 ?>
 
 <div class="site-header-navigation"><div class="site-header-inner">
+	<nav id="site-navigation" class="main-navigation" aria-label="<?php esc_attr_e( 'Primary Menu', 'modern' ); ?>">
 
-<nav id="site-navigation" class="main-navigation" aria-label="<?php esc_attr_e( 'Primary Menu', 'modern' ); ?>">
+		<?php
 
-	<?php if ( $is_mobile_nav_enabled ) : ?>
-	<button id="menu-toggle" class="menu-toggle" aria-controls="menu-primary" aria-expanded="false"><?php echo esc_html_x( 'Menu', 'Mobile navigation toggle button title.', 'modern' ); ?></button>
+		if ( $is_mobile_nav_enabled ) :
+			?>
+			<button
+				id="menu-toggle"
+				class="menu-toggle"
+				aria-controls="menu-primary"
+				aria-expanded="false"
+			><?php
 
-	<?php endif; ?>
-	<div id="site-navigation-container" class="main-navigation-container">
-		<?php wp_nav_menu( Modern_Menu::primary_menu_args( $is_mobile_nav_enabled ) ); ?>
-	</div>
+			echo esc_html_x( 'Menu', 'Mobile navigation toggle button title.', 'modern' );
 
-</nav>
+			?></button>
+			<?php
+		endif;
 
+		?>
+		<div id="site-navigation-container" class="main-navigation-container">
+			<?php
+
+			/**
+			 * For `theme_location` see `Modern_Menu::primary_menu_args()` method
+			 * in `includes/frontend/class-menu.php` file.
+			 */
+			wp_nav_menu( Modern_Menu::primary_menu_args( $is_mobile_nav_enabled ) );
+
+			?>
+		</div>
+
+	</nav>
 </div></div>