+++ template-parts/menu/menu-social.php
@@ -6,39 +6,28 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.2.3
+ * @version  2.5.0
  */
 
+if ( ! has_nav_menu( 'social' ) ) {
+	return;
+}
 
+$cache_key   = Modern_Menu::get_cache_key_social();
+$cache_group = 'modern_' . get_bloginfo( 'language' );
 
+$back_to_top  = '<li class="back-to-top-link">';
+$back_to_top .= '<a href="#top" class="back-to-top animated" title="' . esc_attr__( 'Back to top', 'modern' ) . '">';
+$back_to_top .= '<span class="screen-reader-text">' . esc_html__( 'Back to top &uarr;', 'modern' ) . '</span>';
+$back_to_top .= '</a>';
+$back_to_top .= '</li>';
 
-
-// Requirements check
-
-	if ( ! has_nav_menu( 'social' ) ) {
-		return;
-	}
-
-
-// Variables
-
-	$cache_key   = Modern_Menu::get_cache_key_social();
-	$cache_group = 'modern_' . get_bloginfo( 'language' );
-
-	$back_to_top  = '<li class="back-to-top-link">';
-	$back_to_top .= '<a href="#page" class="back-to-top animated" title="' . esc_attr__( 'Back to top', 'modern' ) . '">';
-	$back_to_top .= '<span class="screen-reader-text">' . esc_html__( 'Back to top &uarr;', 'modern' ) . '</span>';
-	$back_to_top .= '</a>';
-	$back_to_top .= '</li>';
-
-	$social_menu_html = wp_cache_get( $cache_key, $cache_group );
-	$social_menu_args = Modern_Menu::get_menu_args_social( '<ul data-id="%1$s" class="%2$s">%3$s' . $back_to_top . '</ul>' );
-
+$social_menu_html = wp_cache_get( $cache_key, $cache_group );
+$social_menu_args = Modern_Menu::get_menu_args_social( '<ul data-id="%1$s" class="%2$s">%3$s' . $back_to_top . '</ul>' );
 
 ?>
 
 <nav class="social-links" aria-label="<?php esc_attr_e( 'Social Menu', 'modern' ); ?>">
-
 	<?php
 
 	if ( is_customize_preview() ) {
@@ -60,6 +49,10 @@
 		if ( empty( $social_menu_html ) ) {
 			$social_menu_args['echo'] = false;
 
+			/**
+			 * For `theme_location` see `Modern_Menu::get_menu_args_social()` method
+			 * in `includes/frontend/class-menu.php` file.
+			 */
 			$social_menu_html = wp_nav_menu( $social_menu_args );
 			$social_menu_html = str_replace(
 				' id=',
@@ -80,5 +73,4 @@
 	}
 
 	?>
-
 </nav>